requirements.txt
numpy
streamlit
SimpleITK
matplotlib
gitpython
scipy
openai==1.88.0